machine page revision 8 sept
