+++ app/views/home.ejs
@@ -109,6 +109,7 @@
 
     <script>
       function toggleLanguage(language) {
+        // Add variables for more text translations from the element IDs
         let description = document.getElementById("title-aboutus");
         if (language === "Indonesian") {
           description.innerHTML = "Tentang Kami"; // Add more lines for more text translations
