+++ app/views/machinery.ejs
@@ -39,7 +39,7 @@
             src="/image/machine1.png"
             class="d-block w-100"
             alt="..."
-            style="max-height: 92vh"
+            style="max-height: 93vh"
           />
           <div class="carousel-caption d-md-block">
             <div
@@ -63,7 +63,7 @@
             src="/image/slide2machine.png"
             class="d-block w-100"
             alt="..."
-            style="max-height: 92vh"
+            style="max-height: 91vh"
           />
           <div class="carousel-caption d-md-block">
             <div class="title">
