+++ app/views/products.ejs
@@ -161,7 +161,7 @@
         <div class="card">
           <img src="/image/plasticCor.png" />
           <div class="info">
-            <h1 class="h1-card">Military-grade Industrial Plastic Bags</h1>
+            <h1 class="h1-card">Durable Industrial Plastic Bags</h1>
             <p>
               Lorem Ipsum is simply dummy text from the printing and typeseting
               industry
