+++ package.json
@@ -8,7 +8,7 @@
         "express": "^4.18.2"
     },
     "scripts": {
-        "start": "node index.js"
+        "start": "node app.js"
     }
   }
