+++ public/css/home.css
@@ -2,7 +2,7 @@
 @import url("https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;1,100&family=Work+Sans:wght@300&display=swap");
 @import url("https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;1,100&family=Work+Sans:wght@600&display=swap"); */
 @import url("https://fonts.googleapis.com/css2?family=Saira+Condensed:wght@100&display=swap");
-
+@custom-media -- medium (width <= 992px);
 * {
   margin: 0;
   padding: 0;
